DELETE /api/datasets/:id
