+++ flask/app/manage.py
@@ -230,6 +230,14 @@ def add_dummy_data():
                 "condition": "Germline",
                 "input_hpf_path": "",
             },
+            {
+                "tissue_sample_id": 3,
+                "dataset_type": "RGS",
+                "entered": "2020-02-03",
+                "created_by": 1,
+                "condition": "Germline",
+                "input_hpf_path": "",
+            },
         ]
         for d in default_datasets:
             dataset = models.Dataset(
